requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
build==1.2.2.post1
certifi==2024.7.4
charset-normalizer==3.3.0
click==8.1.7
click-default-group==1.2.4
coverage==7.9.2
exceptiongroup==1.3.0
Flask==3.0.0
Flask-Cors==4.0.2
flask-sock==0.7.0
frozenlist==1.4.0
gpt4all==2.8.2
gunicorn==22.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.7
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.5
llm==0.11
llm-gpt4all==0.1.1
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
packaging==23.2
pip-tools==7.4.1
pluggy==1.6.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ulid==2.2.0
PyYAML==6.0.1
requests==2.32.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sqlite-fts4==1.0.3
sqlite-migrate==0.1a2
sqlite-utils==3.35.1
tabulate==0.9.0
tomli==2.2.1
tqdm==4.66.3
typing_extensions==4.8.0
urllib3==2.2.2
Werkzeug==3.0.6
wsproto==1.2.0
yarl==1.12.0
